common page refactor
