+++ src/main/java/com/ycao/mysite/service/markdown/IMarkdownService.java
@@ -4,12 +4,12 @@ import com.ycao.mysite.model.MarkdownFileDomain;
 
 import java.util.List;
 /**
- * used in home/article get all article,
+ * used in index/article get all article,
  * maybe in the future i will ameliorate it by using lazy loading
  * Created by ycao on
  */
 public interface IMarkdownService {
-    public List<MarkdownFileDomain> getAllMarkdown(String userId);
+    public List<MarkdownFileDomain> getAllMarkdown();
     public void deleteOneFile(String mid);
     public void deleteCategoryFiles(String cname,String userId);
 }
