+++ src/main/java/com/ycao/mysite/service/markdown/impl/MarkdownServiceImpl.java
@@ -17,10 +17,8 @@ public class MarkdownServiceImpl implements IMarkdownService {
     MdCatDao mdCatDao;
     //@Cacheable(cacheNames = "markdownCaches",key = "#root.methodName + ':[' + #userId + ']'")
     @Override
-    public List<MarkdownFileDomain> getAllMarkdown(String userId) {
-        if(userId==null)
-            throw BusinessException.withErrorCode(ErrorConstant.Common.PARAM_IS_EMPTY);
-        return mdCatDao.getAllFilesWithCatName(Integer.parseInt(userId));
+    public List<MarkdownFileDomain> getAllMarkdown() {
+        return mdCatDao.getAllFilesWithCatName();
     }
 
     //@CacheEvict(cacheNames = "markdownCaches",allEntries = true,beforeInvocation = true)
